Add write to typesafe config
