+++ examples/src/main/scala/zio/config/examples/ListExample.scala
@@ -4,6 +4,7 @@ import zio.config.ConfigDescriptor._
 import zio.config._
 import zio.config.examples.typesafe.EitherImpureOps
 import zio.config.PropertyTree.Leaf
+import zio.config.typesafe._
 
 // List works quite nicely if the source is typesafe HOCON. Refer typesafe examples
 object ListExample extends App with EitherImpureOps {
@@ -34,8 +35,11 @@ object ListExample extends App with EitherImpureOps {
       )
   )
 
+  val propertyTree =
+    write(config, PgmConfig("something", List("australia", "canada", "usa")))
+
   assert(
-    write(config, PgmConfig("something", List("australia", "canada", "usa"))) ==
+    propertyTree ==
       Right(
         PropertyTree
           .Record(
@@ -46,4 +50,24 @@ object ListExample extends App with EitherImpureOps {
           )
       )
   )
+
+  println(propertyTree.map(_.toHocon))
+  // Right(SimpleConfigObject({"regions":["australia","canada","usa"],"xyz":"something"}))
+  println(propertyTree.map(_.toHocon.render()))
+
+//  Right({
+//    # hardcoded value
+//      "regions" : [
+//    # hardcoded value
+//      "australia",
+//    # hardcoded value
+//      "canada",
+//    # hardcoded value
+//      "usa"
+//    ],
+//    # hardcoded value
+//      "xyz" : "something"
+//  }
+// )
+
 }
